Add fast cache check for remote execution
Hash each part of a depset and cache the resule to increase speed
compared to listing all files, sorting them and creating a merkle tree.

An alias using this fake input root digest is stored in the action
cache by putting the real action key as stdout and stderr digests. When
trying to resolve aliases, the digest is used if stdout is matching
stderr.

This method requires two round trips to the action cache, first to
resolve the alias and then to get the actual result. An improvement
would be to store a copy, but by adding a dummy output file called e.g.
__actual_action_key.digest. Note that the second round trip should still
be done to make sure the "real" action result is not evicted, but that
is no longer on the critical path.