Add in memory buffer to RemoteCache
The in memory buffer in the RemoteCache keeps a content addressable
storage and an action cache. This speeds up builds where the analysis
cache has been discarded.

The action cache makes sure that results are evicted a certian time
after last write. This way, the results will periodically be fetched
from the remote action cache so that the action result and output blobs
won't get evicted.

Having the CAS means that e.g. .d files from C/C++ compilations do not
have to be downloaded from the remote cache. This buffer evicts the
blobs that have not been accessed for a long time. Note that the action
cache eviction strategy makes sure that the blobs are always available
in the remote CAS, so access time expiration is enough.